+++ riverdriver/riverpgxv5/internal/dbsqlc/river_job.sql
@@ -83,6 +83,30 @@ SELECT count(*)
 FROM river_job
 WHERE state = @state;
 
+-- name: JobDelete :one
+WITH job_to_delete AS (
+    SELECT id
+    FROM river_job
+    WHERE river_job.id = @id
+    FOR UPDATE
+),
+deleted_job AS (
+    DELETE
+    FROM river_job
+    USING job_to_delete
+    WHERE river_job.id = job_to_delete.id
+        -- Do not touch running jobs:
+        AND river_job.state != 'running'::river_job_state
+    RETURNING river_job.*
+)
+SELECT *
+FROM river_job
+WHERE id = @id::bigint
+    AND id NOT IN (SELECT id FROM deleted_job)
+UNION
+SELECT *
+FROM deleted_job;
+
 -- name: JobDeleteBefore :one
 WITH deleted_jobs AS (
     DELETE FROM river_job